FixCompilerWarnings
- AbstractConstraint::getType was needlessly returning a ConstraintType as const
- SudokuCell was not using the provided board
- Solver::reduceExactCoverSolutionToSudokuSolution had an unused argument
- XccEleemnt::operator< was needlessly returning a bool as const
- IdPacking utility had unused formal parameters

+++ src/ku/solver/AlgorithmC.cpp
@@ -245,7 +245,7 @@ Forward: {
 
   // Save the currently active sizes
   {
-    if (currentSaveIndex + active + 1 > saveStack.size()) {
+    if (static_cast<size_t>(currentSaveIndex + active + 1) > saveStack.size()) {
       saveStack.resize(currentSaveIndex + active + 1);
     }
     for (int32_t p = 0; p < active; p++) {
